// apps/api/cautela/src/common/enums/cautela-flow-step.enum.ts
export enum CautelaFlowStep {
  SOLICITADA = 'SOLICITADA',
  APROVADA_PELO_GESTOR = 'APROVADA_PELO_GESTOR',
  VALIDADA_PELA_PORTARIA = 'VALIDADA_PELA_PORTARIA',
  SAIDA_AUTORIZADA_PELO_GESTOR = 'SAIDA_AUTORIZADA_PELO_GESTOR',
  ENCERRADA_PELA_PORTARIA = 'ENCERRADA_PELA_PORTARIA',
  REPROVADA = 'REPROVADA',
}